Refactoring (in progress)
