+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectCard/index.tsx
@@ -7,17 +7,20 @@ import { IPersonaSharedProps, Persona, PersonaSize } from 'office-ui-fabric-reac
 import { Icon } from 'office-ui-fabric-react/lib/Icon'
 import strings from 'PortfolioWebPartsStrings'
 import moment from 'moment'
+import { ProjectCardContent } from './ProjectCardContent'
 
 export const ProjectCard: FunctionComponent<IProjectCardProps> = ({
   project,
   actions,
   showProjectOwner,
   showProjectManager,
-  showLifeCycleStatus,
-  showServiceArea,
-  showType,
+  // showLifeCycleStatus,
+  // showServiceArea,
+  // showType,
   phaseLevel
 }) => {
+  
+
   const ownerPersona: IPersonaSharedProps = {
     title: project.owner
       ? `${project.owner.text} | ${strings.ProjectOwner}`
@@ -42,53 +45,53 @@ export const ProjectCard: FunctionComponent<IProjectCardProps> = ({
     }
   }
 
-  const _renderLifeCycleStatus = () => {
-    if (project.lifecycleStatus) {
-      return (
-        <div
-          className={styles.tag}
-          style={
-            project.lifecycleStatus === 'Aktivt'
-              ? { backgroundColor: 'rgb(234,163,0,0.5)', color: 'black' }
-              : { backgroundColor: 'rgb(255,0,0,0.5)', color: 'black' }
-          }>
-          <span>{project.lifecycleStatus}</span>
-        </div>
-      )
-    }
-  }
+  // const _renderLifeCycleStatus = () => {
+  //   if (project.lifecycleStatus) {
+  //     return (
+  //       <div
+  //         className={styles.tag}
+  //         style={
+  //           project.lifecycleStatus === 'Aktivt'
+  //             ? { backgroundColor: 'rgb(234,163,0,0.5)', color: 'black' }
+  //             : { backgroundColor: 'rgb(255,0,0,0.5)', color: 'black' }
+  //         }>
+  //         <span>{project.lifecycleStatus}</span>
+  //       </div>
+  //     )
+  //   }
+  // }
 
-  const _renderServiceAreaText = () => {
-      return (
-        <>
-          {project.serviceArea.map((text, idx) => (
-            <div
-              key={idx}
-              className={styles.tag}
-              style={{ backgroundColor: 'rgb(234,163,0,0.5)', color: 'black' }}>
-              <span>{text}</span>
-            </div>
-          ))}
-        </>
-      )
-  }
+  // const _renderServiceAreaText = () => {
+  //     return (
+  //       <>
+  //         {project.serviceArea.map((text, idx) => (
+  //           <div
+  //             key={idx}
+  //             className={styles.tag}
+  //             style={{ backgroundColor: 'rgb(234,163,0,0.5)', color: 'black' }}>
+  //             <span>{text}</span>
+  //           </div>
+  //         ))}
+  //       </>
+  //     )
+  // }
 
-  const _renderTypeText = () => {
-      return (
-        <>
-          {project.type.map((type, idx) => (
-            <div
-              key={idx}
-              className={styles.tag}
-              style={{ backgroundColor: 'rgb(234,163,0,0.5)', color: 'black' }}>
-              <span>{type}</span>
-            </div>
-          ))}
-        </>
-      )
-  }
+  // const _renderTypeText = () => {
+  //     return (
+  //       <>
+  //         {project.type.map((type, idx) => (
+  //           <div
+  //             key={idx}
+  //             className={styles.tag}
+  //             style={{ backgroundColor: 'rgb(234,163,0,0.5)', color: 'black' }}>
+  //             <span>{type}</span>
+  //           </div>
+  //         ))}
+  //       </>
+  //     )
+  // }
 
-  const endDate = moment(project.endDate).format('DD.MM.YYYY')
+  // const endDate = moment(project.endDate).format('DD.MM.YYYY')
 
   return (
     <a href={project.userIsMember ? project.url : null} style={{ textDecoration: 'none' }}>
@@ -112,7 +115,10 @@ export const ProjectCard: FunctionComponent<IProjectCardProps> = ({
           shouldTruncate={true}
         />
         <hr />
-        <div className={styles.labels}>
+
+        <ProjectCardContent project={project} />
+
+        {/* <div className={styles.labels}>
           {showLifeCycleStatus && _renderLifeCycleStatus()}
           {showServiceArea && _renderServiceAreaText()}
           {showType && _renderTypeText()}
@@ -131,7 +137,7 @@ export const ProjectCard: FunctionComponent<IProjectCardProps> = ({
             />
             <span className={styles.endDateText}>{project.endDate ? endDate : 'Ikke satt'}</span>
           </div>
-        </div>
+        </div> */}
         <div className={styles.footer}>
           <div className={styles.persona}>
             {showProjectOwner && project.owner && (
